+++ packages/core/src/exports/index.ts
@@ -351,6 +351,13 @@ export {
 } from '../actions/watchAccount.js'
 
 export {
+  type WatchAssetParameters,
+  type WatchAssetErrorType,
+  type WatchAssetReturnType,
+  watchAsset,
+} from '../actions/watchAsset.js'
+
+export {
   type WatchBlocksParameters,
   type WatchBlocksReturnType,
   watchBlocks,